Added SubTextures and SpriteSheets

+++ Jah/Source/Core/Application.cpp
@@ -36,8 +36,8 @@ namespace Jah {
 	{
 		while (m_Running)
 		{
-			float time = (float)glfwGetTime();
-			Timestep timestep = time - m_LastFrameTime;
+			double time = glfwGetTime();
+			Timestep timestep = (float)(time - m_LastFrameTime);
 			m_LastFrameTime = time;
 
 			if (!m_Minimized)
